Fix for grass, where newly-compiled grass was invisible.
This happened after Blender 1ce98989e53bae8e4cb7589eb32b043d3e6df05c
in which the behaviour of the object join operator changed: now,
if you join two objects with UV maps with different names, the
resulting object has two UV maps and each one only has half of the
values initialised. So the fix was to rename UV layers on source
objects to be the same thing (standardised on "UVMap").